refactor: parsing 로직 service 계층으로 위임
