+++ src/main/java/lotto/service/LottoService.java
@@ -2,6 +2,7 @@ package lotto.service;
 
 import camp.nextstep.edu.missionutils.Randoms;
 import lotto.domain.Lotto;
+import lotto.domain.WinningNumbers;
 import lotto.dto.PurchasedLottos;
 
 import java.util.ArrayList;
@@ -12,6 +13,7 @@ public class LottoService {
     private static final int MIN_NUMBER = 1;
     private static final int MAX_NUMBER = 45;
     private static final int LOTTO_PRICE = 1000;
+    private static final String DELEMITER = ",";
 
     public PurchasedLottos purchaseLottos(int amount) {
         validateAmount(amount);
@@ -24,6 +26,13 @@ public class LottoService {
         }
 
         return PurchasedLottos.from(lottosNumbers);
+    }
+
+    public WinningNumbers createWinningNumbers(String numbersInput, String bonusInput) {
+        List<Integer> numbers = parseNumbers(numbersInput);
+        int bonusNumber = parseBonusNumber(bonusInput);
+
+        return new WinningNumbers(numbers, bonusNumber);
     }
 
     private void validateAmount(int amount) {
@@ -46,5 +55,28 @@ public class LottoService {
     private Lotto generateLotto() {
         List<Integer> numbers = Randoms.pickUniqueNumbersInRange(MIN_NUMBER, MAX_NUMBER, LOTTO_SIZE);
         return new Lotto(numbers);
+    }
+
+    private List<Integer> parseNumbers(String input) {
+        String[] tokens = input.split(DELEMITER);
+        List<Integer> numbers = new ArrayList<>();
+
+        for (String token : tokens) {
+            try {
+                numbers.add(Integer.parseInt(token.trim()));
+            } catch (NumberFormatException e) {
+                throw new IllegalArgumentException("[ERROR] 당첨 번호는 숫자로 입력해야 합니다.");
+            }
+        }
+
+        return numbers;
+    }
+
+    private int parseBonusNumber(String input) {
+        try {
+            return Integer.parseInt(input.trim());
+        } catch (NumberFormatException e) {
+            throw new IllegalArgumentException("[ERROR] 보너스 번호는 숫자로 입력해야 합니다.");
+        }
     }
 }
